Arquivos atualizados para o funcionamento, segurança básica adicionada

+++ Indioflix/src/main/java/org/Indioflix/Filme.java
@@ -32,7 +32,7 @@ public class Filme {
 	private String diretor;
 	
 	@Column(name = "nota_pessoal")
-	private Integer notaPessoal;
+	private Double notaPessoal;
 	
 	@Column(name = "review_curta", length = 2000)
 	private String reviewCurta;
